@@ -1,8 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
-import './index.css';
 import AppRoutes from './Routes';
+import './index.css';
 
 const App = () => {
 	return (
@@ -12,4 +12,4 @@ const App = () => {
 	);
 };
 
-ReactDOM.render(<App />, document.getElementById('app'));
+ReactDOM.createRoot(document.getElementById('app')).render(<App />);
