@@ -1,10 +1,13 @@
 package com.arbindo.mimock.security.user.models.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.Date;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -21,5 +24,12 @@ public class AuthenticationTokenResponse implements AuthenticationResponse {
             example = "22-06-2022 10:45 PM",
             description = "The expiry timestamp for the token"
     )
-    private String expiresAt;
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
+    private Date expiresAt;
+
+    @Schema(
+            example = "1800",
+            description = "The age of the token in seconds"
+    )
+    private Long expiresAfterSeconds;
 }
