@@ -5,8 +5,8 @@ import { LogError } from './logger/Logger';
 import { config } from '../Config';
 
 const cookies = new Cookies();
-const csrfToken = cookies.get('XSRF-TOKEN');
-const authToken = cookies.get(globalConstants.authCookieName);
+const csrfToken = cookies.get(globalConstants.XSRF_COOKIE_NAME);
+const authToken = () => cookies.get(globalConstants.AUTH_TOKEN_COOKIE_NAME);
 
 export const controller = new AbortController();
 
@@ -23,13 +23,13 @@ const get = async (url, config = {}) => {
 	return await client
 		.get(url, {
 			headers: {
-				Authorization: `Bearer ${authToken}`,
+				Authorization: `Bearer ${authToken()}`,
 			},
 			...config,
 		})
 		.catch((err) => {
 			LogError(err);
-			return err;
+			return Promise.reject(err);
 		});
 };
 
@@ -37,14 +37,14 @@ const post = async (url, requestData, contentType, config = {}) => {
 	return await client
 		.post(url, requestData, {
 			headers: {
-				Authorization: `Bearer ${authToken}`,
+				Authorization: `Bearer ${authToken()}`,
 				'Content-Type': contentType,
 			},
 			...config,
 		})
 		.catch((err) => {
 			LogError(err);
-			return err;
+			return Promise.reject(err);
 		});
 };
 
@@ -58,7 +58,7 @@ const authenticate = async (url, requestData, config = {}) => {
 		})
 		.catch((err) => {
 			LogError(err);
-			return err;
+			return Promise.reject(err);
 		});
 };
 
@@ -66,14 +66,14 @@ const put = async (url, data, contentType, config) => {
 	return await client
 		.put(url, data, {
 			headers: {
-				Authorization: `Bearer ${authToken}`,
+				Authorization: `Bearer ${authToken()}`,
 				'Content-Type': contentType,
 			},
 			...config,
 		})
 		.catch((err) => {
 			LogError(err);
-			return err;
+			return Promise.reject(err);
 		});
 };
 
@@ -81,13 +81,13 @@ const remove = async (url, config) => {
 	return await client
 		.delete(url, {
 			headers: {
-				Authorization: `Bearer ${authToken}`,
+				Authorization: `Bearer ${authToken()}`,
 			},
 			...config,
 		})
 		.catch((err) => {
 			LogError(err);
-			return err;
+			return Promise.reject(err);
 		});
 };
 
@@ -95,13 +95,13 @@ const options = async (url, config) => {
 	return await client
 		.options(url, {
 			headers: {
-				Authorization: `Bearer ${authToken}`,
+				Authorization: `Bearer ${authToken()}`,
 			},
 			...config,
 		})
 		.catch((err) => {
 			LogError(err);
-			return err;
+			return Promise.reject(err);
 		});
 };
 
