@@ -1,3 +1,4 @@
 export const globalConstants = {
-	authCookieName: '__authToken',
+	AUTH_TOKEN_COOKIE_NAME: '__authToken',
+	XSRF_COOKIE_NAME: 'XSRF-TOKEN',
 };
