@@ -50,7 +50,7 @@ module.exports = {
 	plugins: [
 		new HtmlWebPackPlugin({
 			template: './src/index.html',
-			favicon: './src/favicon.ico',
+			favicon: './src/favicon.png',
 		}),
 	],
 };
